Proxies backend requests through API route
Routes requests to the backend through a Next.js API route.

This change removes the direct calls to the backend from the frontend,
and instead uses an API proxy to handle the requests. This allows for
better control over the requests and responses, and allows to manage
the backend URL in a single place.